+++ cmd/main.go
@@ -26,6 +26,7 @@ import (
 	// to ensure that exec-entrypoint and run can make use of them.
 	_ "k8s.io/client-go/plugin/pkg/client/auth"
 
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/runtime"
 	utilruntime "k8s.io/apimachinery/pkg/util/runtime"
 	clientgoscheme "k8s.io/client-go/kubernetes/scheme"
@@ -40,6 +41,7 @@ import (
 	monitoringv1alpha1 "github.com/giantswarm/silence-operator/api/v1alpha1"
 	"github.com/giantswarm/silence-operator/internal/controller"
 	"github.com/giantswarm/silence-operator/pkg/alertmanager"
+	"github.com/giantswarm/silence-operator/pkg/config"
 	// +kubebuilder:scaffold:imports
 )
 
@@ -66,8 +68,8 @@ func main() {
 	var enableHTTP2 bool
 	var tlsOpts []func(*tls.Config)
 
-	var alertmanagerAddress, alertmanagerTenantId string
-	var alertmanagerAuthentication bool
+	var cfg config.Config
+	var silenceSelector string
 	flag.StringVar(&metricsAddr, "metrics-bind-address", "0", "The address the metrics endpoint binds to. "+
 		"Use :8443 for HTTPS or :8080 for HTTP, or leave as 0 to disable the metrics service.")
 	flag.StringVar(&probeAddr, "health-probe-bind-address", ":8081", "The address the probe endpoint binds to.")
@@ -85,15 +87,29 @@ func main() {
 	flag.StringVar(&metricsCertKey, "metrics-cert-key", "tls.key", "The name of the metrics server key file.")
 	flag.BoolVar(&enableHTTP2, "enable-http2", false,
 		"If set, HTTP/2 will be enabled for the metrics and webhook servers")
-	flag.StringVar(&alertmanagerAddress, "alertmanager-address", "http://localhost:9093", "Alertmanager address used to create silences.")
-	flag.StringVar(&alertmanagerTenantId, "alertmanager-default-tenant-id", "", "Alertmanager tenant id.")
-	flag.BoolVar(&alertmanagerAuthentication, "alertmanager-authentication", false, "Enable Alertmanager authentication using Service Account token.")
+	flag.StringVar(&cfg.Address, "alertmanager-address", "http://localhost:9093", "Alertmanager address used to create silences.")
+	flag.StringVar(&cfg.TenantId, "alertmanager-default-tenant-id", "", "Alertmanager tenant id.")
+	flag.BoolVar(&cfg.Authentication, "alertmanager-authentication", false, "Enable Alertmanager authentication using Service Account token.")
+	flag.StringVar(&silenceSelector, "silence-selector", "", "Label selector to filter Silence custom resources (e.g., 'environment=production,tier=frontend').")
 
 	opts := zap.Options{
 		Development: false,
 	}
 	opts.BindFlags(flag.CommandLine)
 	flag.Parse()
+
+	if silenceSelector != "" {
+		parsedSelector, err := metav1.ParseToLabelSelector(silenceSelector)
+		if err != nil {
+			setupLog.Error(err, "unable to parse silence-selector string")
+			os.Exit(1)
+		}
+		cfg.SilenceSelector, err = metav1.LabelSelectorAsSelector(parsedSelector)
+		if err != nil {
+			setupLog.Error(err, "unable to convert silence-selector to labels.Selector")
+			os.Exit(1)
+		}
+	}
 
 	ctrl.SetLogger(zap.New(zap.UseFlagOptions(&opts)))
 
@@ -210,20 +226,14 @@ func main() {
 		os.Exit(1)
 	}
 
+	cfg.BearerToken = mgr.GetConfig().BearerToken
+
 	// TODO Make this more configurable:
-	// - label selector to select silences
 	// - support all alertmanager auth methods
 	// - support multiple tenant (configurable label like observability.giantswarm.io/tenant)
 	var amClient *alertmanager.AlertManager
 	{
-		amConfig := alertmanager.Config{
-			Address:        alertmanagerAddress,
-			Authentication: alertmanagerAuthentication,
-			BearerToken:    mgr.GetConfig().BearerToken,
-			TenantId:       alertmanagerTenantId,
-		}
-
-		amClient, err = alertmanager.New(amConfig)
+		amClient, err = alertmanager.New(cfg)
 		if err != nil {
 			setupLog.Error(err, "unable to setup client", "client", "Alertmanager")
 			os.Exit(1)
@@ -231,9 +241,10 @@ func main() {
 	}
 
 	if err = (&controller.SilenceReconciler{
-		Client:       mgr.GetClient(),
-		Scheme:       mgr.GetScheme(),
-		Alertmanager: amClient,
+		Client:          mgr.GetClient(),
+		Scheme:          mgr.GetScheme(),
+		Alertmanager:    amClient,
+		SilenceSelector: cfg.SilenceSelector,
 	}).SetupWithManager(mgr); err != nil {
 		setupLog.Error(err, "unable to create controller", "controller", "Silence")
 		os.Exit(1)